requirements.txt
whisper
gradio
moviepy
numpy
pandas
torch
transformers
python-dotenv
nltk